requirements.txt
langchain
langchain-core
langchain-community
langchainhub
langgraph
langchain-chroma
sentence-transformers
chromadb
streamlit
python-dotenv
